django/yoga/requirements.txt
appdirs==1.4.3
asgiref==3.2.10
blinker==1.4
CacheControl==0.12.6
certifi==2019.11.28
chardet==3.0.4
click==7.1.2
colorama==0.4.3
contextlib2==0.6.0
deprecation==2.1.0
distlib==0.3.0
distro==1.4.0
Django==3.1.1
Flask==1.1.2
Flask-Admin==1.5.6
Flask-Login==0.5.0
Flask-Mail==0.9.1
html5lib==1.0.1
idna==2.8
ipaddr==2.2.0
itsdangerous==1.1.0
Jinja2==2.11.2
jsonpickle==0.9.6
lockfile==0.12.2
MarkupSafe==1.1.1
msgpack==0.6.2
packaging==20.3
pep517==0.8.2
progress==1.5
psycopg2==2.8.6
pyparsing==2.4.6
python-dateutil==2.8.1
pytoml==0.1.21
pytz==2020.1
requests==2.22.0
retrying==1.3.3
six==1.14.0
sqlparse==0.3.1
squareup==6.3.0.20200826
urllib3==1.25.8
webencodings==0.5.1
Werkzeug==1.0.1
WTForms==2.3.3
